requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiohttp==3.8.1
aiohttp==3.9.3
colorama==0.4.4
discord_py_interactions==5.11.0
discord_py_interactions==5.12.1
discord_py_interactions==4.4.1
google_api_python_client==2.121.0
google_api_python_client==2.129.0
google_auth_oauthlib==1.2.0
interactions_wait_for==1.0.6
protobuf==5.27.2
redis==5.0.2
redis==5.0.4
redis==5.0.1
